// src/api/models/Upgrades.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UpgradeEvent } from './UpgradeEvent';
import {
    UpgradeEventFromJSON,
    UpgradeEventFromJSONTyped,
    UpgradeEventToJSON,
    UpgradeEventToJSONTyped,
} from './UpgradeEvent';

/**
 * 
 * @export
 * @interface Upgrades
 */
export interface Upgrades {
    /**
     * 
     * @type {Array<UpgradeEvent>}
     * @memberof Upgrades
     */
    upgrades: Array<UpgradeEvent>;
}

/**
 * Check if a given object implements the Upgrades interface.
 */
export function instanceOfUpgrades(value: object): value is Upgrades {
    if (!('upgrades' in value) || value['upgrades'] === undefined) return false;
    return true;
}

export function UpgradesFromJSON(json: any): Upgrades {
    return UpgradesFromJSONTyped(json, false);
}

export function UpgradesFromJSONTyped(json: any, ignoreDiscriminator: boolean): Upgrades {
    if (json == null) {
        return json;
    }
    return {
        
        'upgrades': ((json['upgrades'] as Array<any>).map(UpgradeEventFromJSON)),
    };
}

export function UpgradesToJSON(json: any): Upgrades {
    return UpgradesToJSONTyped(json, false);
}

export function UpgradesToJSONTyped(value?: Upgrades | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'upgrades': ((value['upgrades'] as Array<any>).map(UpgradeEventToJSON)),
    };
}