// src/api/models/APM.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface APM
 */
export interface APM {
    /**
     * 
     * @type {string}
     * @memberof APM
     */
    playerName: string;
    /**
     * 
     * @type {number}
     * @memberof APM
     */
    actionCount: number;
    /**
     * 
     * @type {number}
     * @memberof APM
     */
    minutes: number;
    /**
     * 
     * @type {number}
     * @memberof APM
     */
    apm: number;
}

/**
 * Check if a given object implements the APM interface.
 */
export function instanceOfAPM(value: object): value is APM {
    if (!('playerName' in value) || value['playerName'] === undefined) return false;
    if (!('actionCount' in value) || value['actionCount'] === undefined) return false;
    if (!('minutes' in value) || value['minutes'] === undefined) return false;
    if (!('apm' in value) || value['apm'] === undefined) return false;
    return true;
}

export function APMFromJSON(json: any): APM {
    return APMFromJSONTyped(json, false);
}

export function APMFromJSONTyped(json: any, ignoreDiscriminator: boolean): APM {
    if (json == null) {
        return json;
    }
    return {
        
        'playerName': json['playerName'],
        'actionCount': json['actionCount'],
        'minutes': json['minutes'],
        'apm': json['apm'],
    };
}

export function APMToJSON(json: any): APM {
    return APMToJSONTyped(json, false);
}

export function APMToJSONTyped(value?: APM | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'playerName': value['playerName'],
        'actionCount': value['actionCount'],
        'minutes': value['minutes'],
        'apm': value['apm'],
    };
}