// src/api/models/FirstBlood.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FirstBlood
 */
export interface FirstBlood {
    /**
     * 
     * @type {string}
     * @memberof FirstBlood
     */
    attacker: string;
    /**
     * 
     * @type {string}
     * @memberof FirstBlood
     */
    victim: string;
    /**
     * 
     * @type {number}
     * @memberof FirstBlood
     */
    atMinute: number;
}

/**
 * Check if a given object implements the FirstBlood interface.
 */
export function instanceOfFirstBlood(value: object): value is FirstBlood {
    if (!('attacker' in value) || value['attacker'] === undefined) return false;
    if (!('victim' in value) || value['victim'] === undefined) return false;
    if (!('atMinute' in value) || value['atMinute'] === undefined) return false;
    return true;
}

export function FirstBloodFromJSON(json: any): FirstBlood {
    return FirstBloodFromJSONTyped(json, false);
}

export function FirstBloodFromJSONTyped(json: any, ignoreDiscriminator: boolean): FirstBlood {
    if (json == null) {
        return json;
    }
    return {
        
        'attacker': json['attacker'],
        'victim': json['victim'],
        'atMinute': json['atMinute'],
    };
}

export function FirstBloodToJSON(json: any): FirstBlood {
    return FirstBloodToJSONTyped(json, false);
}

export function FirstBloodToJSONTyped(value?: FirstBlood | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'attacker': value['attacker'],
        'victim': value['victim'],
        'atMinute': value['atMinute'],
    };
}